Update ShoppingListDeleteController.php
If the shopping list is already deleted an exception is thrown in twig because the idShoppingList is null and the route to delete the shopping list can not get generated. 
To avoid this fail whale we want to redirect the customer back to the shopping list page. Unfortunately it is not possible without creating a new route, controller, .... because the Return Type of the deleteConfirmAction is fixed to View and does not allow a RedirectResponse. 

For that reason we suggest to remove the return type of this method.